+++ src/components/editor/markdown-preview/markdown-it-plugins/parser-debugger.ts
@@ -1,0 +1,8 @@
+import MarkdownIt from 'markdown-it/lib'
+
+export const MarkdownItParserDebugger: MarkdownIt.PluginSimple = (md: MarkdownIt) => {
+  md.core.ruler.push('test', (state) => {
+    console.log(state)
+    return true
+  })
+}