+++ src/components/editor/markdown-preview/regex-plugins/replace-legacy-speakerdeck-short-code.ts
@@ -1,0 +1,11 @@
+import { RegexOptions } from '../../../../external-types/markdown-it-regex/interface'
+
+const finalRegex = /^{%speakerdeck (\w+\/[\w-]+) ?%}$/
+
+export const replaceLegacySpeakerdeckShortCode: RegexOptions = {
+  name: 'legacy-speakerdeck-short-code',
+  regex: finalRegex,
+  replace: (match) => {
+    return `<a target="_blank" rel="noopener noreferrer" href="https://speakerdeck.com//${match}">https://speakerdeck.com/${match}</a>`
+  }
+}