+++ src/components/editor/markdown-preview/regex-plugins/replace-legacy-youtube-short-code.ts
@@ -1,0 +1,11 @@
+import { RegexOptions } from '../../../../external-types/markdown-it-regex/interface'
+
+export const replaceLegacyYoutubeShortCode: RegexOptions = {
+  name: 'legacy-youtube-short-code',
+  regex: /^{%youtube ([^"&?\\/\s]{11}) ?%}$/,
+  replace: (match) => {
+    // ESLint wants to collapse this tag, but then the tag won't be valid html anymore.
+    // noinspection CheckTagEmptyBody
+    return `<codimd-youtube id="${match}"></codimd-youtube>`
+  }
+}