+++ src/components/editor/markdown-preview/replace-components/video-util.ts
@@ -1,0 +1,12 @@
+import { DomElement } from 'domhandler'
+
+export const getIdFromCodiMdTag = (node: DomElement, tagName: string): (string | undefined) => {
+  if (node.name !== `codimd-${tagName}` || !node.attribs || !node.attribs.id) {
+    return
+  }
+  return node.attribs.id
+}
+
+export interface VideoFrameProps {
+  id: string
+}