+++ src/external-types/markdown-it-emoji/index.d.ts
@@ -1,0 +1,6 @@
+
+declare module 'markdown-it-emoji' {
+  import MarkdownIt from 'markdown-it/lib'
+  const markdownItEmoji: MarkdownIt.PluginSimple
+  export = markdownItEmoji
+}