+++ src/external-types/markdown-it-regex/index.d.ts
@@ -1,0 +1,6 @@
+declare module 'markdown-it-regex' {
+  import MarkdownIt from 'markdown-it/lib'
+  import { RegexOptions } from './interface'
+  const markdownItRegex: MarkdownIt.PluginWithOptions<RegexOptions>
+  export = markdownItRegex
+}