+++ src/external-types/markdown-it-regex/interface.ts
@@ -1,0 +1,5 @@
+export interface RegexOptions {
+  name: string,
+  regex: RegExp,
+  replace: (match: string) => string
+}