+++ src/external-types/markdown-it-task-lists/index.d.ts
@@ -1,0 +1,6 @@
+
+declare module 'markdown-it-task-lists' {
+  import MarkdownIt from 'markdown-it/lib'
+  const markdownItTaskLists: MarkdownIt.PluginSimple
+  export = markdownItTaskLists
+}